pyproject.toml
[project]
name = "ai-tutor-for-github-repositories"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "qdrant-client>=1.16.2",
    "tiktoken>=0.12.0",
    "sentence-transformers>=2.2.0",
    "groq>=1.0.0",
    "langgraph>=1.0.5",
    "langchain>=1.2.0",
    "langchain-core>=1.2.5",
    "redis>=5.0.0",
    "tenacity>=8.2.0",
    "docker>=7.0.0",
    "websockets>=15.0.1",
    "black>=25.12.0",
    "ruff>=0.14.13",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning:google.protobuf.internal.well_known_types",
    "ignore::FutureWarning:keras.src.export.tf2onnx_lib",
]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | \.ruff_cache
  | __pycache__
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults (FastAPI Depends is correct)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "(\\.venv|build|dist|__pycache__)/",
]
